+++ pysubgroup/__init__.py
@@ -3,8 +3,9 @@ from pysubgroup.algorithms import *
 from pysubgroup.measures import *
 from pysubgroup.utils import *
 
-from pysubgroup.boolean_target import *
+from pysubgroup.nominal_target import *
 from pysubgroup.numeric_target import *
 from pysubgroup.fi_target import *
 
 from pysubgroup.visualization import *
+from pysubgroup.complex_target import *
